Quality of life changes made to ingestor
I noticed that there is some difference in the response of the code when using the scopus or science direct data. I think it is best to split the the two searches more completely, so I have created two Ingestion engines, one for scopus and one for science direct. I will work on scopus first, and maybe one of the students can modify the code for science direct.
This also means that I need to center my code on scopus for now.